// src/app/models/seat-statistics.ts
export class SeatStatistics {
    totalCapacityCount?: number;
    onSaleCount?: number;

    pendingCount?: number;
    soldCount?: number;
    blockedCount?: number;
    canceledCount?: number;
    selectedCount?: number;

    soldIndividualCount?: number;
    soldSeasonalCount?: number;
    soldCompCount?: number;
    soldGroupCount?: number;

    reservedIndividualCount?: number;
    reservedCompCount?: number;
}